bed/requirements.txt
absl-py==1.0.0
aiosqlite==0.17.0
anyio==3.4.0
appdirs==1.4.4
argcomplete==1.12.3
argon2-cffi==21.1.0
asgiref==3.4.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bandit==1.7.1
black==21.12b0
bleach==4.1.0
btrfsutil==5.15.1
CacheControl==0.12.6
cachetools==4.2.4
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
contextlib2==0.6.0.post1
crit==0.0.1
cryptography==36.0.1
cycler==0.11.0
databases==0.5.3
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.3
distro==1.6.0
dnspython==2.1.0
entrypoints==0.3
fastapi==0.70.0
flatbuffers==2.0
fonttools==4.28.3
gast==0.5.3
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.43.0
h11==0.12.0
h5py==3.6.0
html5lib==1.1
idna==3.3
importlib-metadata==4.8.1
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isc==2.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyter-packaging==0.11.1
jupyter-server==1.13.1
jupyterlab==3.2.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.2
keras==2.8.0rc0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
libfdt==1.6.1
libvirt-python==7.8.0
lit==13.0.0.dev0
lockfile==0.12.2
lxml==4.6.4
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
meson==0.60.3
mistune==0.8.4
more-itertools==8.10.0
msgpack==1.0.3
multitasking==0.0.10
mypy-extensions==0.4.3
nbclassic==0.3.4
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.4
nftables==0.1
notebook==6.4.4
numpy==1.21.4
oauthlib==3.1.1
OpenCC==1.1.2
opt-einsum==3.3.0
ordered-set==4.0.2
packaging==21.3
pandas==1.3.4
pandas-datareader==0.10.0
pandocfilters==1.4.3
parso==0.8.3
pathspec==0.9.0
pbr==5.8.0
pep517==0.12.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pipx==0.16.4
platformdirs==2.4.0
ply==3.11
podman-compose==1.0.3
progress==1.6
prometheus-client==0.11.0
prompt-toolkit==3.0.24
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
pulsemixer==1.5.1
py-bcrypt==0.4
pyalsa==1.2.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pydantic==1.8.2
pydocstyle==6.1.1
Pygments==2.10.0
PyGObject==3.42.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
PySide2==5.15.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.2.1
rados==2.0.0
rbd==2.0.0
Reflector==2021.7.8.19.52.52
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.3.0
resolvelib==0.5.5
retrying==1.3.3
rgw==2.0.0
rsa==4.8
scikit-learn==1.0.1
scipy==1.7.3
Send2Trash==1.8.0
shiboken2==5.15.2
shtab==1.5.2
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==2.2.0
SQLAlchemy==1.4.28
sqlalchemy2-stubs==0.0.2a19
sqlparse==0.4.2
SSSDConfig==2.6.1
starlette==0.16.0
stevedore==3.5.0
team==1.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.8.0rc0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.11.1
testpath==0.5.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.0.0
tldr==3.0.0
toml==0.10.2
tomli==1.2.2
tomlkit==0.8.0
tornado==6.1
traitlets==5.1.1
typing-extensions==4.0.1
tzdata==2021.5
urllib3==1.26.7
userpath==1.7.0
uvicorn==0.16.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.59.0
websockets==10.1
Werkzeug==2.0.2
widgetsnbextension==3.5.1
wrapt==1.13.3
yfinance==0.1.67
zipp==3.6.0
